+++ packages/cli/src/builders/repo-builder.ts
@@ -1,27 +1,31 @@
 import {
   ClassDeclaration,
+  ConstructorDeclaration,
   printNode,
   Project,
   QuoteKind,
+  Scope,
   SourceFile,
   StructureKind,
+  SyntaxKind,
   VariableDeclarationKind,
 } from 'ts-morph';
 import { CallExpression, factory, ScriptTarget } from 'typescript';
-import { RequestsCacheBuilder } from './requests-cache.builder';
-import { ActiveIdsBuilder } from './active-ids.builder';
-import { EntitiesBuilder } from './entities.builder';
-import { UIEntitiesBuilder } from './ui-entities.builder';
-import { RequestsStatusBuilder } from './requests-status.builder';
-import { ActiveIdBuilder } from './active-id.builder';
-import { PropsBuilder } from './props.builder';
 import { Options } from '../types';
 import { names, resolveStoreVariableName } from '../utils';
+import { ActiveIdBuilder } from './active-id.builder';
+import { ActiveIdsBuilder } from './active-ids.builder';
+import { EntitiesBuilder } from './entities.builder';
+import { PropsBuilder } from './props.builder';
+import { RequestsCacheBuilder } from './requests-cache.builder';
+import { RequestsStatusBuilder } from './requests-status.builder';
+import { UIEntitiesBuilder } from './ui-entities.builder';
 
 export function createRepo(options: Options) {
   const { storeName } = options;
   const storeNames = names(storeName);
   const isFunctionTpl = !options.template || options.template === 'functions';
+  const isStoreInlinedInClass = !isFunctionTpl && options.inlineStoreInClass;
 
   const project = new Project({
     manipulationSettings: {
@@ -39,10 +43,16 @@ export function createRepo(options: Options) {
     name: repoName,
     isExported: true,
   });
+  let repoClassDecConstructor: ConstructorDeclaration | undefined;
+
+  if (isStoreInlinedInClass) {
+    repoClassDec.addConstructor();
+    repoClassDecConstructor = repoClassDec.getConstructors()[0];
+  }
 
   sourceFile.addImportDeclaration({
     moduleSpecifier: '@ngneat/elf',
-    namedImports: ['Store', 'createState'].map((name) => ({
+    namedImports: ['createStore'].map((name) => ({
       kind: StructureKind.ImportSpecifier,
       name,
     })),
@@ -70,37 +80,37 @@ export function createRepo(options: Options) {
     }
   }
 
-  const state = factory.createCallExpression(
-    factory.createIdentifier('createState'),
-    undefined,
-    propsFactories
+  const storeOpts = factory.createIdentifier(
+    `{ name: '${storeNames.propertyName}' }`
   );
 
-  const repoPosition = repoClassDec.getChildIndex();
+  const store = factory.createCallExpression(
+    factory.createIdentifier('createStore'),
+    undefined,
+    [storeOpts, ...propsFactories]
+  );
 
-  sourceFile.insertVariableStatement(repoPosition, {
-    declarationKind: VariableDeclarationKind.Const,
-    isExported: isFunctionTpl,
-    declarations: [
-      {
-        name: resolveStoreVariableName(options.template, storeNames),
-        initializer: `new Store({ name: '${storeNames.propertyName}', state, config })`,
-      },
-    ],
-  });
+  if (isStoreInlinedInClass && repoClassDecConstructor) {
+    addInlineStoreToRepoClass({
+      repoClassDec,
+      repoClassDecConstructor,
+      options,
+      store,
+      storeNames,
+    });
+  } else {
+    addStoreToRepo({
+      repoClassDec,
+      options,
+      store,
+      storeNames,
+      isFunctionTpl,
+      sourceFile,
+    });
 
-  sourceFile.insertVariableStatement(repoPosition, {
-    declarationKind: VariableDeclarationKind.Const,
-    declarations: [
-      {
-        name: '{ state, config }',
-        initializer: printNode(state),
-      },
-    ],
-  });
-
-  if (isFunctionTpl) {
-    toFunctions(sourceFile, repoClassDec);
+    if (isFunctionTpl) {
+      toFunctions(sourceFile, repoClassDec);
+    }
   }
 
   if (options.hooks) {
@@ -110,6 +120,86 @@ export function createRepo(options: Options) {
   sourceFile.formatText({ indentSize: 2 });
 
   return sourceFile.getText();
+}
+
+function addStoreToRepo({
+  repoClassDec: classDec,
+  sourceFile,
+  options,
+  storeNames,
+  store,
+  isFunctionTpl,
+}: {
+  repoClassDec: ClassDeclaration;
+  sourceFile: SourceFile;
+  options: Options;
+  storeNames: ReturnType<typeof names>;
+  store: CallExpression;
+  isFunctionTpl: boolean;
+}) {
+  const repoPosition = classDec.getChildIndex();
+
+  sourceFile.insertVariableStatement(repoPosition, {
+    declarationKind: VariableDeclarationKind.Const,
+    isExported: isFunctionTpl,
+    declarations: [
+      {
+        name: resolveStoreVariableName(options.template, storeNames),
+        initializer: printNode(store),
+      },
+    ],
+  });
+}
+
+function addInlineStoreToRepoClass({
+  repoClassDec: classDec,
+  repoClassDecConstructor: constructorDec,
+  options,
+  storeNames,
+  store,
+}: {
+  repoClassDec: ClassDeclaration;
+  repoClassDecConstructor: ConstructorDeclaration;
+  options: Options;
+  storeNames: ReturnType<typeof names>;
+  store: CallExpression;
+}) {
+  const storeName = resolveStoreVariableName(
+    options.template,
+    storeNames,
+    true
+  );
+  const { propertyIndex, methodIndex } = getPositionsOfInlineStoreDeclarations(
+    classDec,
+    constructorDec
+  );
+
+  const createStoreMethodName = 'createStore';
+
+  classDec.insertMethod(methodIndex, {
+    name: createStoreMethodName,
+    returnType: `typeof store`,
+    scope: Scope.Private,
+    statements: (writer) => {
+      writer.writeLine(`const store = ${printNode(store)};`);
+      writer.blankLine();
+      writer.writeLine(`return store;`);
+    },
+  });
+
+  constructorDec.insertStatements(
+    0,
+    `${storeName} = this.${createStoreMethodName}();`
+  );
+
+  const storeProperty = classDec.insertProperty(propertyIndex, {
+    name: `${resolveStoreVariableName(options.template, storeNames)}`,
+    scope: Scope.Private,
+  });
+
+  if (propertyIndex > 0) {
+    storeProperty?.prependWhitespace('\n');
+  }
 }
 
 function toFunctions(sourceFile: SourceFile, classDec: ClassDeclaration) {
@@ -128,4 +218,23 @@ function toFunctions(sourceFile: SourceFile, classDec: ClassDeclaration) {
   sourceFile.replaceWithText(
     `${sourceFile.getText()}\n ${exported.join('\n\n')}`
   );
+}
+
+function getPositionsOfInlineStoreDeclarations(
+  classDec: ClassDeclaration,
+  constructorDec: ConstructorDeclaration
+) {
+  const lastPropertyIndex = classDec
+    .getLastChildByKind(SyntaxKind.PropertyDeclaration)
+    ?.getChildIndex();
+  const lastMethodIndex = classDec
+    .getLastChildByKind(SyntaxKind.MethodDeclaration)
+    ?.getChildIndex();
+
+  return {
+    methodIndex: (lastMethodIndex ?? constructorDec.getChildIndex()) + 1,
+    propertyIndex: lastPropertyIndex
+      ? lastPropertyIndex + 1
+      : constructorDec.getChildIndex(),
+  };
 }