+++ packages/cli/src/specs/repo-builder.spec.ts
@@ -15,6 +15,20 @@ describe('Repo Builder', () => {
     expect(output).toMatchSnapshot('withProps');
   });
 
+  it('withProps in a class', () => {
+    const output = createRepo({
+      idKey: 'id',
+      features: ['withProps'],
+      crud: [],
+      template: 'class',
+      inlineStoreInClass: true,
+      storeName: 'todos',
+      path: '',
+    });
+
+    expect(output).toMatchSnapshot('withProps in a class');
+  });
+
   it('withRequestsCache', () => {
     const output = createRepo({
       idKey: 'id',
@@ -26,6 +40,20 @@ describe('Repo Builder', () => {
     });
 
     expect(output).toMatchSnapshot('withRequestsCache');
+  });
+
+  it('withRequestsCache in a class', () => {
+    const output = createRepo({
+      idKey: 'id',
+      features: ['withRequestsCache'],
+      crud: [],
+      template: 'class',
+      inlineStoreInClass: true,
+      storeName: 'todos',
+      path: '',
+    });
+
+    expect(output).toMatchSnapshot('withRequestsCache in a class');
   });
 
   it('withRequestsStatus', () => {
@@ -41,6 +69,20 @@ describe('Repo Builder', () => {
     expect(output).toMatchSnapshot('withRequestsStatus');
   });
 
+  it('withRequestsStatus in a class', () => {
+    const output = createRepo({
+      idKey: 'id',
+      features: ['withRequestsStatus'],
+      crud: [],
+      template: 'class',
+      inlineStoreInClass: true,
+      storeName: 'todos',
+      path: '',
+    });
+
+    expect(output).toMatchSnapshot('withRequestsStatus in a class');
+  });
+
   it('withActiveId', () => {
     const output = createRepo({
       idKey: 'id',
@@ -52,6 +94,20 @@ describe('Repo Builder', () => {
     });
 
     expect(output).toMatchSnapshot('withActiveId');
+  });
+
+  it('withActiveId in a class', () => {
+    const output = createRepo({
+      idKey: 'id',
+      features: ['withActiveId'],
+      crud: [],
+      template: 'class',
+      inlineStoreInClass: true,
+      storeName: 'todos',
+      path: '',
+    });
+
+    expect(output).toMatchSnapshot('withActiveId in a class');
   });
 
   it('withActiveIds', () => {
@@ -67,6 +123,20 @@ describe('Repo Builder', () => {
     expect(output).toMatchSnapshot('withActiveIds');
   });
 
+  it('withActiveIds in a class', () => {
+    const output = createRepo({
+      idKey: 'id',
+      features: ['withActiveIds'],
+      crud: [],
+      template: 'class',
+      inlineStoreInClass: true,
+      storeName: 'todos',
+      path: '',
+    });
+
+    expect(output).toMatchSnapshot('withActiveIds in a class');
+  });
+
   it('withEntities', () => {
     const output = createRepo({
       idKey: 'id',
@@ -78,6 +148,20 @@ describe('Repo Builder', () => {
     });
 
     expect(output).toMatchSnapshot('withEntities');
+  });
+
+  it('withEntities in a class', () => {
+    const output = createRepo({
+      idKey: 'id',
+      features: ['withEntities'],
+      crud: [],
+      template: 'class',
+      inlineStoreInClass: true,
+      storeName: 'todos',
+      path: '',
+    });
+
+    expect(output).toMatchSnapshot('withEntities in a class');
   });
 
   it('withEntities idKey', () => {
@@ -93,6 +177,20 @@ describe('Repo Builder', () => {
     expect(output).toMatchSnapshot('withEntities');
   });
 
+  it('withEntities idKey in a class', () => {
+    const output = createRepo({
+      idKey: '_id',
+      features: ['withEntities'],
+      crud: ['addEntities', 'updateEntities', 'deleteEntities'],
+      template: 'class',
+      inlineStoreInClass: true,
+      storeName: 'todos',
+      path: '',
+    });
+
+    expect(output).toMatchSnapshot('withEntities in a class');
+  });
+
   it('withUIEntities', () => {
     const output = createRepo({
       idKey: 'id',
@@ -106,6 +204,20 @@ describe('Repo Builder', () => {
     expect(output).toMatchSnapshot('withUIEntities');
   });
 
+  it('withUIEntities in a class', () => {
+    const output = createRepo({
+      idKey: 'id',
+      features: ['withUIEntities'],
+      crud: [],
+      template: 'class',
+      inlineStoreInClass: true,
+      storeName: 'todos',
+      path: '',
+    });
+
+    expect(output).toMatchSnapshot('withUIEntities in a class');
+  });
+
   it('all', () => {
     const output = createRepo({
       idKey: 'id',
@@ -117,6 +229,20 @@ describe('Repo Builder', () => {
     });
 
     expect(output).toMatchSnapshot('all');
+  });
+
+  it('all in a class', () => {
+    const output = createRepo({
+      idKey: 'id',
+      features: baseFeatures.map(({ value }) => value),
+      crud: ['deleteEntities', 'addEntities', 'updateEntities'],
+      template: 'class',
+      inlineStoreInClass: true,
+      storeName: 'todos',
+      path: '',
+    });
+
+    expect(output).toMatchSnapshot('all in a class');
   });
 
   it('should support function template', () => {