+++ packages/cli/src/utils.ts
@@ -10,9 +10,12 @@ export function coerceArray<T>(value: T[] | T): T[] {
 
 export function resolveStoreVariableName(
   template: Options['template'],
-  { propertyName }: ReturnType<typeof names>
+  { propertyName }: ReturnType<typeof names>,
+  inlineStoreInClass: boolean = false
 ) {
-  return template === 'functions' ? `${propertyName}Store` : 'store';
+  return template === 'functions'
+    ? `${propertyName}Store`
+    : `${inlineStoreInClass ? 'this.' : ''}store`;
 }
 
 /**