+++ packages/persist-state/src/lib/persist-state.ts
@@ -35,21 +35,23 @@ export function persistState<S extends Store>(store: S, options: Options<S>) {
   const { setItem, getItem } = options.storage;
   const initialized = new ReplaySubject<boolean>(1);
 
-  from(getItem(merged.key!)).subscribe((value) => {
-    if (value) {
-      store.update((state) => {
-        return merged.preStoreInit!({
-          ...state,
-          ...value,
+  const loadFromStorageSubscription = from(getItem(merged.key!)).subscribe(
+    (value) => {
+      if (value) {
+        store.update((state) => {
+          return merged.preStoreInit!({
+            ...state,
+            ...value,
+          });
         });
-      });
+      }
+
+      initialized.next(true);
+      initialized.complete();
     }
+  );
 
-    initialized.next(true);
-    initialized.complete();
-  });
-
-  const subscription = merged.source!(store)
+  const saveToStorageSubscription = merged.source!(store)
     .pipe(
       skip(1),
       switchMap((value) => setItem(merged.key!, value))
@@ -59,7 +61,8 @@ export function persistState<S extends Store>(store: S, options: Options<S>) {
   return {
     initialized$: initialized.asObservable(),
     unsubscribe() {
-      subscription.unsubscribe();
+      saveToStorageSubscription.unsubscribe();
+      loadFromStorageSubscription.unsubscribe();
     },
   };
 }