+++ packages/entities/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+# [3.3.0](https://github.com/ngneat/elf/compare/entities-3.2.1...entities-3.3.0) (2022-02-23)
+
+
+### Features
+
+* **store:** 🔥 introduce createStore function ([#174](https://github.com/ngneat/elf/issues/174)) ([fbc68ff](https://github.com/ngneat/elf/commit/fbc68ff1dd91190486a74dea9637ce34a47fb9ea))
+
+
+
+## [3.2.1](https://github.com/ngneat/elf/compare/entities-3.2.0...entities-3.2.1) (2022-02-12)
+
+
+### Bug Fixes
+
+* **entities:** 🐞 export update by ids ([8a3de8d](https://github.com/ngneat/elf/commit/8a3de8d9868a8adbac7f3c2cbe98ed0310239a7a))
+
+
+
 # [3.2.0](https://github.com/ngneat/elf/compare/entities-3.1.1...entities-3.2.0) (2022-02-12)