+++ packages/persist-state/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.1.1](https://github.com/ngneat/elf/compare/persist-state-1.1.0...persist-state-1.1.1) (2022-02-16)
+
+
+### Bug Fixes
+
+* **persist-state:** unsubscribe from storage ([#156](https://github.com/ngneat/elf/issues/156)) ([5d7e9a4](https://github.com/ngneat/elf/commit/5d7e9a449631637f506de14e618838cfc0c6c4ec))
+
+
+
 # [1.1.0](https://github.com/ngneat/elf/compare/persist-state-1.0.0...persist-state-1.1.0) (2022-02-06)