+++ packages/store/CHANGELOG.md
@@ -2,6 +2,60 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.3.3](https://github.com/ngneat/elf/compare/store-1.3.2...store-1.3.3) (2022-02-24)
+
+
+### Bug Fixes
+
+* **store:** 🐞 fix createStore type ([f1f38b2](https://github.com/ngneat/elf/commit/f1f38b26157ca237c1cb60f67ce055bab627fef7))
+
+
+
+## [1.3.2](https://github.com/ngneat/elf/compare/store-1.3.1...store-1.3.2) (2022-02-23)
+
+
+### Bug Fixes
+
+* **store:** 🐞 revert type ([370bb9c](https://github.com/ngneat/elf/commit/370bb9cfb33b98dd88058469297e070ddaf2f83d))
+
+
+
+## [1.3.1](https://github.com/ngneat/elf/compare/store-1.3.0...store-1.3.1) (2022-02-23)
+
+
+### Bug Fixes
+
+* **store:** 🐞 fix createStore type ([87c756b](https://github.com/ngneat/elf/commit/87c756b79c47cc0d225f3bc4633ae50a3363ef9e))
+
+
+
+# [1.3.0](https://github.com/ngneat/elf/compare/store-1.2.0...store-1.3.0) (2022-02-23)
+
+
+### Features
+
+* **store:** 🔥 introduce createStore function ([#174](https://github.com/ngneat/elf/issues/174)) ([fbc68ff](https://github.com/ngneat/elf/commit/fbc68ff1dd91190486a74dea9637ce34a47fb9ea))
+
+
+
+# [1.2.0](https://github.com/ngneat/elf/compare/store-1.1.0...store-1.2.0) (2022-02-17)
+
+
+### Features
+
+* **store:** 🔥 add setProps mutation ([452d2aa](https://github.com/ngneat/elf/commit/452d2aa56df7097aff4b387eee97702b0b1c4f24))
+
+
+
+# [1.1.0](https://github.com/ngneat/elf/compare/store-1.0.2...store-1.1.0) (2022-02-16)
+
+
+### Features
+
+* **store:** 🔥 add setProp mutation ([b0bea61](https://github.com/ngneat/elf/commit/b0bea61e67e476559a58ce35018460f8bfac68fb))
+
+
+
 ## [1.0.2](https://github.com/ngneat/elf/compare/store-1.0.1...store-1.0.2) (2021-12-14)